Refactor: Outsource action mapping envs/
