+++ nudge/agents/utils_atari.py
@@ -237,45 +237,3 @@ def state_to_extended_repr(state, coin_jump, swap_coins=None, coin0_x=None):
     tr_ext = [*er, *er_add]
 
     return tr_ext, swap_coins, coin0_x
-
-
-def preds_to_action_atari(action, prednames):
-    """
-    map explaining to action
-    0:noop
-    1:up
-    2:down
-
-    CJA_MOVE_LEFT: Final[int] = 1
-    CJA_MOVE_RIGHT: Final[int] = 2
-    CJA_MOVE_UP: Final[int] = 3
-    """
-    if len(prednames) == 13: # asterix:
-        if 'noop' in prednames[action]:
-            return 0
-        elif 'up' in prednames[action]:
-            return 1
-        elif 'right' in prednames[action]:
-            return 2
-        elif 'left' in prednames[action]:
-            return 3
-        elif 'down' in prednames[action]:
-            return 4
-    else: # freeway
-        if 'noop' in prednames[action]:
-            return 0
-        elif 'up' in prednames[action]:
-            return 1
-        elif 'down' in prednames[action]:
-            return 2
-
-
-def action_map_atari(prediction, args, prednames=None):
-    """map model action to game action"""
-    if args.alg == 'ppo':
-        # simplified action--- only left right up
-        # action = coin_jump_actions_from_unified(torch.argmax(predictions).cpu().item() + 1)
-        action = prediction + 1
-    elif args.alg == 'logic':
-        action = preds_to_action_atari(prediction, prednames)
-    return action
