+++ nudge/collect_data.py
@@ -10,8 +10,9 @@ from environments.getout.getout.getout.getout import Getout
 from environments.getout.getout.getout.paramLevelGenerator import ParameterizedLevelGenerator
 from envs.getout.state_extraction import sample_to_model_input, extract_state, collate
 from agents.neural_agent import ActorCritic
-from agents.utils_loot import extract_neural_state_loot, simplify_action_loot, extract_logic_state_loot
-from agents.utils_threefish import extract_logic_state_threefish, extract_neural_state_threefish
+from agents.utils_loot import extract_neural_state_loot, extract_logic_state_loot
+from envs.loot.actions import simplify_action_loot
+from envs.threefish.state_extraction import extract_logic_state as extract_logic_state_threefish, extract_neural_state as extract_neural_state_threefish
 from tqdm import tqdm
 from nsfr.utils.common import extract_for_cgen_explaining
